get: Use a shared object to store return values for metadata
Each time the process was returning it copied the data_all var which
even though each one will add metadata it won't affect the version that we
write back out to data_all.json. Resulting in that file missing the
datagetter block.

Also make sure on any exception we set valid and downloads to false