requirements: Unpin very old importlib-metadata
It's unclear that importlib-metadata needs to be pinned in
`requirements.in`, as it's a core package we don't use directly.
This pin of a very old version is causing issues with trying to use
newer versions of libraries in TSG/datastore.

It appears that no other ODSC or TSG project pins this package in their
`requirements.in` files, so I think this unpin should be safe.